requirements.txt
customtkinter
packaging
requests
pyinstaller
pytest
python-dotenv
ctktable